Add logic to reader frontend and api backend

+++ api-backend/controllers/commentController.js
@@ -18,14 +18,19 @@ const getCommentsForPost = async (req, res) => {
 
 // Create a new comment
 const createComment = async (req, res) => {
-	const { content, username, email, postId } = req.body;
+	const { content, postId } = req.body;
+	const userId = req.user.userId;
+
+	const user = await prisma.user.findUnique({
+		where: { id: userId },
+	});
 
 	try {
 		const newComment = await prisma.comment.create({
 			data: {
 				content,
-				username,
-				email,
+				username: user.username,
+				email: user.email,
 				postId,
 			},
 		});
@@ -38,12 +43,30 @@ const createComment = async (req, res) => {
 // Edit comment
 const updateComment = async (req, res) => {
 	const { id } = req.params;
-	const { content, username, email } = req.body;
+	const { content } = req.body;
+	const userId = req.user.userId;
 
 	try {
+		const comment = await prisma.comment.findUnique({
+			where: { id: parseInt(id) },
+		});
+
+		if (!comment) {
+			return res.status(404).json({ error: "Comment not found" });
+		}
+
+		const user = await prisma.user.findUnique({
+			where: { id: userId },
+		});
+
+		// Only allow editing if the username/email match
+		if (comment.email !== user.email) {
+			return res.status(401).json({ error: "You are not authorized to edit this comment." });
+		}
+
 		const updated = await prisma.comment.update({
 			where: { id: parseInt(id) },
-			data: { content, username, email },
+			data: { content },
 		});
 		res.json(updated);
 	} catch (error) {
@@ -54,8 +77,26 @@ const updateComment = async (req, res) => {
 // Delete comment
 const deleteComment = async (req, res) => {
 	const { id } = req.params;
+	const userId = req.user.userId;
 
 	try {
+		const comment = await prisma.comment.findUnique({
+			where: { id: parseInt(id) },
+		});
+
+		if (!comment) {
+			return res.status(404).json({ error: "Comment not found" });
+		}
+
+		const user = await prisma.user.findUnique({
+			where: { id: userId },
+		});
+
+		// Only allow deletion if the username/email match
+		if (comment.email !== user.email) {
+			return res.status(403).json({ error: "You are not authorized to delete this comment." });
+		}
+
 		const deleted = await prisma.comment.delete({
 			where: { id: parseInt(id) },
 		});
